定义 repository 接口，便于支持更多的数据库实现
