+++ src/main/kotlin/io/zhudy/duic/repository/UserRepository.kt
@@ -15,114 +15,45 @@
  */
 package io.zhudy.duic.repository
 
-import com.mongodb.client.model.Filters.eq
-import com.mongodb.client.model.IndexModel
-import com.mongodb.client.model.IndexOptions
-import com.mongodb.client.model.Projections.exclude
-import com.mongodb.client.model.Projections.include
-import com.mongodb.client.model.Updates.combine
-import com.mongodb.client.model.Updates.set
-import com.mongodb.reactivestreams.client.MongoDatabase
 import io.zhudy.duic.domain.Page
 import io.zhudy.duic.domain.Pageable
 import io.zhudy.duic.domain.User
-import org.bson.Document
-import org.bson.types.ObjectId
-import org.springframework.stereotype.Repository
+import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
-import reactor.core.publisher.toFlux
-import reactor.core.publisher.toMono
-import reactor.core.scheduler.Schedulers
-import java.util.*
 
 /**
  * @author Kevin Zou (kevinz@weghst.com)
  */
-@Repository
-class UserRepository(
-        private val mongo: MongoDatabase
-) {
-
-    companion object {
-        private val USER_COLL_NAME = "user"
-    }
-
-    init {
-        mongo.getCollection(USER_COLL_NAME).createIndexes(listOf(
-                IndexModel(Document("email", 1), IndexOptions().background(true).unique(true))
-        )).toMono().subscribe()
-    }
+interface UserRepository {
 
     /**
      *
      */
-    fun insert(user: User) = mongo.getCollection(USER_COLL_NAME).insertOne(Document(
-            mapOf(
-                    "_id" to ObjectId().toString(),
-                    "email" to user.email,
-                    "password" to user.password,
-                    "created_at" to Date(),
-                    "updated_at" to Date()
-            )
-    )).toMono()
+    fun insert(user: User): Mono<*>
 
     /**
      *
      */
-    fun delete(email: String) = mongo.getCollection(USER_COLL_NAME).deleteOne(eq("email", email))
-            .toMono()
-            .map { it.deletedCount.toInt() }
+    fun delete(email: String): Mono<Int>
 
     /**
      *
      */
-    fun updatePassword(email: String, password: String) = mongo.getCollection(USER_COLL_NAME).updateOne(
-            eq("email", email),
-            combine(
-                    set("password", password),
-                    set("updated_at", Date())
-            ))
-            .toMono()
-            .map { it.modifiedCount.toInt() }
+    fun updatePassword(email: String, password: String): Mono<Int>
 
     /**
      *
      */
-    fun findByEmail(email: String): Mono<User> = mongo.getCollection(USER_COLL_NAME).find(eq("email", email))
-            .first()
-            .toMono()
-            .map {
-                User(
-                        email = it.getString("email"),
-                        password = it.getString("password")
-                )
-            }
-
-    fun findPage(pageable: Pageable) = Mono.zip(
-            mongo.getCollection(USER_COLL_NAME).find()
-                    .projection(exclude("password"))
-                    .skip(pageable.offset)
-                    .limit(pageable.size)
-                    .toFlux()
-                    .publishOn(Schedulers.parallel())
-                    .map {
-                        User(
-                                email = it.getString("email"),
-                                createdAt = it.getDate("created_at"),
-                                updatedAt = it.getDate("updated_at")
-                        )
-                    }
-                    .collectList(),
-            mongo.getCollection(USER_COLL_NAME).count().toMono()
-    ).map {
-        Page(it.t1, it.t2.toInt(), pageable)
-    }
+    fun findByEmail(email: String): Mono<User>
 
     /**
      *
      */
-    fun findAllEmail() = mongo.getCollection(USER_COLL_NAME).find()
-            .projection(include("email"))
-            .toFlux()
-            .map { it.getString("email") }
+    fun findPage(pageable: Pageable): Mono<Page<User>>
+
+    /**
+     *
+     */
+    fun findAllEmail(): Flux<String>
+
 }
